Solved supplier outputting null supplier ID
Problem occured because Supplier.supplierID was an unintialized String object. The Supplier constructor did not have any functionality to set the ID for the supplier. Updated so it now accepts three arguments (id, name, and address) instead of just two (name and address).

The tester, Warehouse, and UserInterface classes were changed to reflect this.